+++ src/main/java/at/qe/sepm/skeleton/ui/controllers/TeamDetailController.java
@@ -9,6 +9,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
 
+import javax.faces.application.FacesMessage;
+import javax.faces.context.FacesContext;
 import java.io.Serializable;
 import java.util.List;
 
@@ -55,10 +57,47 @@ public class TeamDetailController implements Serializable {
 
 
     public void doDeleteTeam(){
-        this.teamService.deleteTeam(team);
-        team = null;
+        FacesMessage message;
+        if (checkIfDeletionIsAllowed(team)) {
+            try {
+                this.teamService.deleteTeam(team);
+                team = null;
+            }
+            catch(Exception e) {
+                e.printStackTrace();
+            }
+        }
+        else {
+            System.out.println("hello");
+            warnMessage("team deletion", "You can't delete this team");
+            return;
+        }
+
+        successMessage("team deletion", "Team deleted");
+
     }
 
+    public boolean checkIfDeletionIsAllowed (Team team){
+        if (!userService.getUsersOfTeam(team).isEmpty()) {
+            return false;
+        }
+        else if (userService.getTeamLeader(team) != null) {
+            return false;
+        }
+        return true;
+    }
+
+    public static void warnMessage(String target, String message) {
+        addMessage(target, new FacesMessage(FacesMessage.SEVERITY_WARN, "Warning!", message));
+    }
+
+    public static void successMessage(String target, String message) {
+        addMessage(target, new FacesMessage(FacesMessage.SEVERITY_INFO, "Success!", message));
+    }
+
+    private static void addMessage(String target, FacesMessage message) {
+        FacesContext.getCurrentInstance().addMessage(target, message);
+    }
 
     public User getEmployee() {
         return employeeAdd;
